# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dragon Math API",
    "version": "1.0.0",
    "description": "The routes the Dragon Academy iOS app uses. Generated from server/contracts/ by `npm run openapi`; do not edit by hand."
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "user"
        ]
      },
      "User": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/ChildUser"
          },
          {
            "$ref": "#/components/schemas/AdultUser"
          }
        ],
        "discriminator": {
          "propertyName": "account_type",
          "mapping": {
            "child": "#/components/schemas/ChildUser",
            "parent": "#/components/schemas/AdultUser",
            "admin": "#/components/schemas/AdultUser"
          }
        }
      },
      "ChildUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "account_type": {
            "type": "string",
            "enum": [
              "child"
            ]
          },
          "current_node_id": {
            "type": "integer"
          },
          "avatar": {
            "type": "string"
          },
          "font": {
            "type": "string"
          },
          "dragon_trial_completed": {
            "type": "boolean"
          },
          "needs_handle": {
            "type": "boolean"
          },
          "effective_plan": {
            "type": "string",
            "description": "Highest plan among the child's guardians: free, premium or classroom."
          },
          "entitlements": {
            "$ref": "#/components/schemas/Entitlements"
          },
          "family_mode": {
            "type": "boolean",
            "description": "Present and true when the session came from a shared family device."
          }
        },
        "required": [
          "id",
          "username",
          "account_type",
          "current_node_id",
          "avatar",
          "font",
          "dragon_trial_completed",
          "needs_handle",
          "effective_plan",
          "entitlements"
        ]
      },
      "Entitlements": {
        "type": "object",
        "properties": {
          "games_locked": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Game ids (src/data/games.js) the child cannot open on their plan."
          }
        },
        "required": [
          "games_locked"
        ]
      },
      "AdultUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "account_type": {
            "type": "string",
            "enum": [
              "parent",
              "admin"
            ]
          },
          "email": {
            "type": "string",
            "nullable": true,
            "description": "The login email. For a Sign in with Apple account it may be a private relay address."
          },
          "email_verified": {
            "type": "boolean"
          },
          "contact_email": {
            "type": "string",
            "nullable": true,
            "description": "Where progress digests and COPPA notices go. Null until the parent sets one."
          },
          "contact_email_verified": {
            "type": "boolean"
          },
          "adult_role": {
            "type": "string",
            "description": "parent or teacher."
          },
          "plan": {
            "type": "string",
            "description": "free, premium or classroom."
          }
        },
        "required": [
          "id",
          "username",
          "account_type",
          "email",
          "email_verified",
          "contact_email",
          "contact_email_verified",
          "adult_role",
          "plan"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable message, safe to show a kid or parent."
          }
        },
        "required": [
          "error"
        ]
      },
...
      "DragonCollectionResponse": {
        "type": "object",
        "properties": {
          "owned": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OwnedDragon"
            }
          },
          "catalog": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogDragon"
            },
            "description": "Every non-retired dragon, in dragon_id order."
          },
          "total_dragons": {
            "type": "integer",
            "description": "Length of catalog, for the \"X / total collected\" headline."
          }
        },
        "required": [
          "owned",
          "catalog",
          "total_dragons"
        ]
      },
      "OwnedDragon": {
        "type": "object",
        "properties": {
          "dragon_id": {
            "type": "integer"
          },
          "count": {
            "type": "integer",
            "description": "How many of this dragon the child has caught."
          },
          "first_acquired_at": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 timestamp of the first catch."
          },
          "name": {
            "type": "string",
            "nullable": true,
            "description": "Null when the dragon is not in the catalog."
          },
          "rarity": {
            "type": "string",
            "description": "common, uncommon, rare, very_rare, legendary or mythic."
          }
        },
        "required": [
          "dragon_id",
          "count",
          "first_acquired_at",
          "name",
          "rarity"
        ]
      },
      "CatalogDragon": {
        "type": "object",
        "properties": {
          "dragon_id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "rarity": {
            "type": "string",
            "description": "common, uncommon, rare, very_rare, legendary or mythic."
          }
        },
        "required": [
          "dragon_id",
          "name",
          "rarity"
        ]
      },
...
      "MemoryPassagesResponse": {
        "type": "object",
        "properties": {
          "passages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoryPassage"
            }
          }
        },
        "required": [
          "passages"
        ]
      },
      "MemoryPassage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "description": "What kind of text it is, verse, poem, quote, speech, definition or other."
          },
          "body": {
            "type": "string",
            "description": "The text to learn, exactly as the grown-up entered it."
          },
          "mastery_level": {
            "type": "integer",
            "description": "Hardest level practised: 0 none, 1 easy, 2 medium, 3 hard."
          },
          "last_practiced_at": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 timestamp, or null if never practised."
          },
          "created_at": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 timestamp."
          },
          "updated_at": {
            "type": "string",
            "description": "ISO 8601 timestamp of the last edit; identifies the revision a practice session was of."
          }
        },
        "required": [
          "id",
          "title",
          "category",
          "body",
          "mastery_level",
          "last_practiced_at",
          "created_at",
          "updated_at"
        ]
      },
...
    },
...
  },
  "paths": {
    "/api/auth/me": {
      "get": {
        "operationId": "getCurrentUser",
        "summary": "The signed-in user.",
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The signed-in user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dragons": {
      "get": {
        "operationId": "getDragonCollection",
        "summary": "The signed-in child's dragon collection, with the active catalog.",
        "tags": [
          "dragons"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "The collection.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DragonCollectionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/memory-passages": {
      "get": {
        "operationId": "listMemoryPassages",
        "summary": "A child's memorize passages, oldest first.",
        "tags": [
          "memorize"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "exclusiveMinimum": true,
              "description": "Required for a grown-up (a linked child's id). A child may omit it, or pass their own id."
            },
            "required": false,
            "description": "Required for a grown-up (a linked child's id). A child may omit it, or pass their own id.",
            "name": "child_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "The passages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryPassagesResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request was invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, malformed or expired credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Signed in, but not allowed to do this.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}